+++ projects/demo/src/app/app.module.ts
@@ -2,18 +2,16 @@ import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
-import {
-  MatButtonModule,
-  MatCardModule,
-  MatChipsModule,
-  MatFormFieldModule,
-  MatInputModule,
-  MatToolbarModule,
-  MatSelectModule,
-  MatSlideToggleModule,
-  MatTooltipModule,
-} from '@angular/material';
-import { LoggerModule, NGXLogger, NgxLoggerLevel } from 'ngx-logger';
+import { MatButtonModule } from '@angular/material/button';
+import { MatCardModule } from '@angular/material/card';
+import { MatChipsModule } from '@angular/material/chips';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatInputModule } from '@angular/material/input';
+import { MatSelectModule } from '@angular/material/select';
+import { MatSlideToggleModule } from '@angular/material/slide-toggle';
+import { MatToolbarModule } from '@angular/material/toolbar';
+import { MatTooltipModule } from '@angular/material/tooltip';
+import { LoggerModule, NgxLoggerLevel } from 'ngx-logger';
 
 import { AppComponent } from './app.component';
 import { LogConfigComponent } from './log-config/log-config.component';