+++ CHANGELOG.md
@@ -7,7 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - Merged features waiting to be published in upcoming version
 
+### Fixed
+
 - Fix getcolor, colorscheme was not binded correctly [#253](https://github.com/dbfannin/ngx-logger/pull/253). Thanks [@jschank](https://github.com/jschank)
+
+### Added
+
+- Updates to Angular 11 [#251](https://github.com/dbfannin/ngx-logger/pull/251). Thanks [@bmtheo](https://github.com/bmtheo)
 
 
 ## [4.2.1] - 2021-02-23